hotel-service integrated with distributed cache and kafka facilities

+++ hotel-service/src/main/java/io/hotely/hotel/controllers/exceptions/CityNotFoundException.java
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class CityNotFoundException extends RuntimeException {
-  public CityNotFoundException(Long id) {
+  public CityNotFoundException(Integer id) {
     super("Could not find city with id: " + id);
   }
 }
